+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 ---------
 
+v3.1.0
+------
+- A new high-performant, optimized implementation for the flat-combining synchronization technique. This new implementation provided by the Synch framework is written from the scratch.
+- A few API fixes.
+
 v3.0.1
 ------
 - README.md updated.